+++ Models/JwtTokenResponse.cs
@@ -4,8 +4,6 @@ namespace garge_operator.Models
 {
     public class JwtTokenResponse
     {
-        [JsonPropertyName("$id")]
-        public required string Id { get; set; }
         [JsonPropertyName("token")]
         public required string Token { get; set; }
     }